repo cleanup and added readme
