+++ .gitignore
@@ -65,3 +65,5 @@ node_modules/
 # dotenv environment variables file
 .env
 
+#macOS
+.DS_Store
