+++ functions/src/index.ts
@@ -1,39 +1,28 @@
-import * as functions from "firebase-functions";
-const authFunctions = require('./auth/auth');
-const jwt = require("express-jwt");
-const jwksRsa = require("jwks-rsa");
-const express = require("express");
-const cors = require("cors");
-
-const app = express();
-
-// Automatically allow cross-origin requests
-app.use(cors({ origin: true }));
-
-
 /**
- * Only handle routes and connecting them to functions here
+ * Handle express routing in this file
  */
 
-const checkJwt = jwt({
-  secret: jwksRsa.expressJwtSecret({
-    cache: true,
-    rateLimit: true,
-    jwksRequestsPerMinute: 5,
-    jwksUri: `https://${functions.config().auth0.domain}/.well-known/jwks.json`,
-  }),
+import * as functions from "firebase-functions";
+const authFunctions = require("./auth/auth");
+import app from './express';
 
-  audience: functions.config().auth0.audience,
-  issuer: `https://${functions.config().auth0.domain}/`,
-  algorithm: ["RS256"],
+//this will match every call made to this api.
+app.all((request: any, response: any, next: any) => {
+  // check if the user has access to relevant permissions. If not then deny access
+  // this function may end up pretty complex and need to be moved into a separate file under ./auth/verifyPermissions.ts
+  // an alternative simpler way is to ensure that only requests that are validated on the front-end come through
+  // if there is no way to call the api through the front end then is this additional check required?
+  // we will likely have a separate auth0 production client id to ensure that just cloning & running the repo will not suffice
+  if (false) {
+    response.send("Unauthorized! Access Denied");
+  }
+
+  //next() basically says to run the next route that matches the url
+  next();
 });
 
-// Add middleware to authenticate requests
+app.get("/getCustomToken", authFunctions.getCustomToken);
+app.post("/createTestUser", authFunctions.createTestUser);
 
-app.get('/getCustomToken',checkJwt, authFunctions.createCustomToken);
-app.post('/createTestUser', authFunctions.createTestUser);
-app.get('/hello-world', (req: any, res: any) => {
-    res.send("hello world!");
-})
 
 exports.api = functions.https.onRequest(app);
